chore: move to hook
